pyproject.toml
[project]
name = "audiology-api"
version = "0.1.0"
description = "A CDK project for analyzing audiological reports using LLMs."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aws-cdk-lib==2.200.1",
    "aws-lambda-powertools>=3.15.1",
    "constructs>=10.0.0,<11.0.0",
    "ipdb>=0.13.13",
    "jaraco-test>=5.5.1",
    "pytest>=6.2.5",
    "websocket>=0.2.1",
    "websocket-client>=1.8.0",
    "websockets>=15.0.1",
    "zipp>=3.23.0",
]

[dependency-groups]
api = [
    "aws-lambda-powertools>=3.15.1",
    "boto3>=1.38.40",
]
dev = [
    "boto3>=1.38.40",
    "botocore>=1.38.40",
    "pytest==6.2.5",
]
